requirements.txt
PySide6>=6.8.0
markdown>=3.5.1
python-docx>=0.8.11
reportlab>=3.6.0
weasyprint>=59.0
html2text>=2020.1.16
